notification_service/requirements.txt
flask
requests
secure-smtplib
python-dotenv
